+++ examples/erp/drizzlify/config.ts
@@ -1,6 +1,7 @@
 import z from 'zod/v4'
 
 import { defineNextJsServerConfig } from '@genseki/next'
+import { phone } from '@genseki/plugins'
 import { defineServerConfig } from '@genseki/react'
 
 import { foodsCollection } from './collections/foods'
@@ -8,6 +9,28 @@ import { usersCollection } from './collections/users'
 import { baseConfig, builder } from './helper'
 
 const baseServerConfig = defineServerConfig(baseConfig, {
+  plugins: [
+    phone({
+      //  TODO: fix relation type
+      baseConfig,
+      sendOtp: async (phone) => {
+        console.log(`Sending OTP to phone: ${phone}`)
+        return {
+          status: 'success',
+          token: '1234567890abcdef',
+          refno: 'ref123456',
+        }
+      },
+      verifyOtp: async (args: { token: string; pin: string }) => {
+        console.log(`Verifying OTP: ${args.pin} for token: ${args.token}`)
+        return true // Simulate successful verification
+      },
+      signUpOnVerification: {
+        getTempEmail: (phoneNumber) => `${phoneNumber.replace(/\D/g, '')}@example.com`,
+        getTempName: (phoneNumber) => `User ${phoneNumber.replace(/\D/g, '')}`,
+      },
+    }),
+  ],
   collections: {
     users: usersCollection,
     foods: foodsCollection,