+++ packages/plugins/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @genseki/plugins
 
+## 0.1.0-alpha.11
+
+## 0.1.0-alpha.10
+
+## 0.1.0-alpha.9
+
+## 0.1.0-alpha.8
+
+### Minor Changes
+
+- [#67](https://github.com/softnetics/genseki/pull/67) [`49927f2`](https://github.com/softnetics/genseki/commit/49927f2f7593ad4c43895789948291b02e52f2bb) Thanks [@saenyakorn](https://github.com/saenyakorn)! - Add Phone plugins
+
 ## 0.1.0-alpha.7
 
 ### Patch Changes